% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocument.R
\docType{data}
\name{similarity}
\alias{similarity}
\title{Simirarity or Distance between two points.}
\format{An object of class \code{matrix} with 20 rows and 20 columns.}
\usage{
similarity
}
\description{
A dataset containing a matrix of similarity between two points. 20 row and 20 col.
}
\keyword{datasets}

