% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hmds}
\alias{hmds}
\title{Calcurate approximate positions in high dimensions}
\usage{
hmds(data, dim = 50L, approx = 1.01, itera = 1000L)
}
\arguments{
\item{data}{symmetric NumericMatrix of input data. it describe the distance between two points.}

\item{dim}{int: it describe dimentions of distance space between two poitns.}

\item{approx}{double: A limit of distance of all pairs points in the space.}

\item{itera}{int: iteration numbers to search space between two points.}
}
\value{
Numeric Matrix of points in a distance space. row is items. col is dimentions.
}
\description{
Calcurate approximate positions in high dimensions
}
\examples{
output <- hmds(matrix(10, nrow=3, ncol=3),dim=20,approx=1.2,itera=10000);
}

